+++ Prototipo/messaggi/Mission.h
@@ -28,7 +28,7 @@ public:
 	int getD_LRGB(){return D_LRGB;};
 	void setD_STOP(int D){ D_STOP=D;};
 	int getD_STOP(){return D_STOP;};
-	void setQ_DOORS( FermataType Q){ Q_DOORS=Q;};
+	void setQ_DOORS( FermataType Q){ Q_DOORS=(int)Q;};
 	int getQ_DOORS(){return Q_DOORS;};
 	void setT_DOORS_TIME( int T){ T_DOORS_TIME=T;};
 	int getT_DOORS_TIME(){return T_DOORS_TIME;};
